# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for image management for your backend.\nMade with Go, Fiber and MongoDB",
        "title": "Saint Ark: File Uploading",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "eam.dayan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/images": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image-handler"
                ],
                "summary": "Get all images from the server.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "specify file format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "0000-00-00",
                        "description": "date",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "get the status of server.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "image-handler"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "type": "file",
                        "description": "single file upload",
                        "name": "myFile",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    }
}